+++ src/drto/__init__.py
@@ -38,6 +38,7 @@ from drto import dynamic_simulation as _dynamic_simulation  # noqa: F401
 from drto import dynamic_to_steady_state as _dynamic_to_steady_state  # noqa: F401
 from drto import infinite_horizon as _infinite_horizon  # noqa: F401
 from drto import parameterize as _parameterize  # noqa: F401
+from drto import steady_state_optimization as _steady_state_optimization  # noqa: F401
 from drto import steady_state_simulation as _steady_state_simulation  # noqa: F401
 
 try: